Big update for backend
views.py/models.py
-break down views structure to multiple file
- all test crud/api of every model except post and postRequest
- using cloud to store image(cloudinary)

urls.py
- called views functions from there own app's urls.py instead of having every function at project's urls.py
