+++ Backend/admins/admin.py
@@ -1,3 +1,8 @@
 from django.contrib import admin
+from .models import Admin, Post, PostRequest
+
 
 # Register your models here.
+admin.site.register(Admin)
+admin.site.register(Post)
+admin.site.register(PostRequest)
