+++ Backend/admins/models.py
@@ -18,11 +18,11 @@ class Post(models.Model):
     adminID = models.ForeignKey(to = Admin, on_delete=models.CASCADE)
     itemID = models.ForeignKey(to = Item, on_delete=models.CASCADE)
     isActive = models.BooleanField(default=True)
-    datePost = models.DateField()
+    datePost = models.DateTimeField()
     
     def save(self, *args, **kwargs):
         ''' On save, update timestamps '''
-        if not self.id:
+        if not self.postID:
             self.created = timezone.now()
         self.modified = timezone.now()
         return super(Post, self).save(*args, **kwargs)
