+++ Backend/users/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import User, Category, SubCategory, Place, Item
+from .models import User, Category, SubCategory, Place, Item, Request
 
 # Register your models here.
 
@@ -8,3 +8,4 @@ admin.site.register(Category)
 admin.site.register(SubCategory)
 admin.site.register(Place)
 admin.site.register(Item)
+admin.site.register(Request)
