+++ Backend/users/views.py/__init__.py
@@ -1,0 +1,7 @@
+from .index import *
+from .users import *
+from .categories import *
+from .subcategories import *
+from .places import *
+from .items import *
+from .requests import *
