+++ Backend/users/views.py/index.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def index(request):
+    return render(request, 'index.html')
